+++ app.ts
@@ -1,18 +1,4 @@
-import axios from "axios";
-const fs = require("fs");
-
-interface Url {
-  value: string;
-}
-console.log("object");
-interface localFile {
-  input: string,
-  output: string
-}
-
-const URL: Url = {
-  value: process.env.SOURCE_FILE,
-} 
+/* import axios from "axios";
 
 const LOCALFILE: localFile = {
   input: "./tmp/publicDebateReports.zip",
@@ -34,28 +20,4 @@ const downloadZipFile = new Promise((resolve, reject) => {
     data.pipe(fs.createWriteStream(LOCALFILE.input));  
   }); */
 
-const zlib = require('zlib');
-
-var yauzl = require("yauzl");
-
-yauzl.open(LOCALFILE.input, {lazyEntries: true}, function(err:any, zipfile:any) {
-  if (err) throw err;
-  zipfile.readEntry();
-  zipfile.on("entry", function(entry:any) {
-    if (/\/$/.test(entry.fileName)) {
-      // Directory file names end with '/'.
-      // Note that entries for directories themselves are optional.
-      // An entry's fileName implicitly requires its parent directories to exist.
-      zipfile.readEntry();
-    } else {
-      // file entry
-      zipfile.openReadStream(entry, function(err:any, readStream:any) {
-        if (err) throw err;
-        readStream.on("end", function() {
-          zipfile.readEntry();
-        });
-        readStream.pipe(fs.createWriteStream(LOCALFILE.output));
-      });
-    }
-  });
-});
+
