+++ utils/constants.ts
@@ -1,0 +1,13 @@
+/* interface Url {
+  value: string;
+}
+
+interface localFile {
+  input: string;
+  output: string;
+}
+
+const URL: Url = {
+  value: process.env.SOURCE_FILE,
+};
+ */
